% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_MIC.R
\name{estimate_MIC}
\alias{estimate_MIC}
\title{Estimate MIC}
\usage{
estimate_MIC(data, covars, response)
}
\arguments{
\item{data}{data.frame containing response as one column, and various covariates as other columns}

\item{covars}{vector of potential covariate names, which match column names in \code{data}}

\item{response}{name of response, which matches one column name in \code{data}}
}
\description{
Estimate various MIC-related statistics
}
\author{
Kevin See
}
