% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_partial_dependence.R
\name{plot_partial_dependence}
\alias{plot_partial_dependence}
\title{Partial Dependence Plots}
\usage{
plot_partial_dependence(
  rf_mod,
  data,
  plot_covars = NULL,
  data_dict = NULL,
  type = c("quantile", "mean"),
  pred_quantile = 0.9,
  n_pts = 200,
  log_transform = T,
  log_offset = 0.005,
  scales = "free"
)
}
\arguments{
\item{rf_mod}{the \code{quantregForest} model}

\item{data}{the data.frame used to fit the \code{rf_mod}}

\item{plot_covars}{character vector of covariates to plot}

\item{data_dict}{data.frame containing columns \code{}}

\item{type}{whether to predict a particular quantile, or the mean}

\item{pred_quantile}{if \code{type} is "quantile", which quantile to predict, between 0 and 1? The default value is 0.9}

\item{n_pts}{how many points to use in predictions?}

\item{log_transform}{was the response log transformed?}

\item{log_offset}{if an offset was used before log transforming, enter it here.}

\item{scales}{ggplot facet_wrap argument \code{scales} for facetting}

\item{...}{other arguements to be passed to \code{ggplot}}
}
\value{
ggplot object
}
\description{
Plot the partial dependence of the response for each covariate
}
\author{
Kevin See
}
